Fix SQLite JDBC URL construction
Prior to this change, SQLite JDBC URLs where constructed using the java
URL class. The output file would be converted to a string and fed into
the URL's constructor as a "scheme specific part". This only works,
however, if the File does not contain any path information. URL escapes
certain characters required for file paths on Windows, resulting in a
file created in the current working directory with a mangled name.

To fix this issue, string concatentation is used instead of the URL
class to prevent character escaping.